+++ fresh.gen.ts
@@ -8,7 +8,9 @@ import * as $api_joke from "./routes/api/joke.ts";
 import * as $greet_name_ from "./routes/greet/[name].tsx";
 import * as $index from "./routes/index.tsx";
 import * as $login from "./routes/login.tsx";
+import * as $test1 from "./routes/test1.tsx";
 import * as $Counter from "./islands/Counter.tsx";
+import * as $test1_1 from "./islands/test1.tsx";
 import { type Manifest } from "$fresh/server.ts";
 
 const manifest = {
@@ -19,9 +21,11 @@ const manifest = {
     "./routes/greet/[name].tsx": $greet_name_,
     "./routes/index.tsx": $index,
     "./routes/login.tsx": $login,
+    "./routes/test1.tsx": $test1,
   },
   islands: {
     "./islands/Counter.tsx": $Counter,
+    "./islands/test1.tsx": $test1_1,
   },
   baseUrl: import.meta.url,
 } satisfies Manifest;
