+++ routes/test1.tsx
@@ -1,21 +1,9 @@
 import type { Signal } from "@preact/signals";
 import { Button } from "../components/Button.tsx";
+import Testisland from "../islands/test1.tsx"
 
-export function Test(){
-    const handleClick = () => {
-        // GoogleのURLにリダイレクト
-        window.location.href = "https://google.co.jp/";
-    };
-    return(
-        <div class="px-4 py-8 mx-auto bg-[#86efac]">
-            <h1>welcome to test page</h1>
-            <div class="flex gap-8 py-6">
-                <Button onClick={handleClick}>waa</Button>
-
-            </div>
-
-
-
-        </div>
-    );
+export default function Test(){
+    return (<div>
+        <Testisland></Testisland>
+    </div>)
 }
